Moved Replay class to it's own file
